hr_management/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic>=2.7.4
pydantic-settings==2.1.0

# Database
pymongo==4.6.1
motor==3.3.2

# AI
google-generativeai>=0.3.0

# RAG - Updated versions
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
langchain-core>=1.0.0
chromadb>=0.4.22

# PDF
pdfplumber==0.10.3
python-multipart==0.0.6
